refactor: create Config.scala
