+++ engine/language-server/src/main/scala/org/enso/languageserver/data/Environment.scala
@@ -1,28 +1,5 @@
 package org.enso.languageserver.data
 
-import java.io.File
-import java.util.UUID
-
-import org.enso.languageserver.filemanager.{
-  ContentRootNotFound,
-  FileSystemFailure
-}
-
-/**
-  * The config of the running Language Server instance.
-  *
-  * @param contentRoots a mapping between content root id and absolute path to
-  *                     the content root
-  */
-case class Config(contentRoots: Map[UUID, File] = Map.empty) {
-
-  def findContentRoot(rootId: UUID): Either[FileSystemFailure, File] =
-    contentRoots
-      .get(rootId)
-      .toRight(ContentRootNotFound)
-
-}
-
 /**
   * The state of the running Language Server instance.
   *
